Organize repository to conform to a conventional JavaScript project
