--- bin.js
@@ -1,6 +1,0 @@
-#!/usr/bin/env node
-
-const TerminalTextEditor = require('./index.js');
-
-
-let text_editor = new TerminalTextEditor();
