--- denpendency.sh
@@ -1,2 +1,0 @@
-npm install
-
