--- install-linux.sh
@@ -1,4 +1,0 @@
-apt-get update
-apt-get install curl
-curl -fsSL https://deb.nodesource.com/setup_14.x | bash
-apt-get install nodejs
