--- run.sh
@@ -1,5 +1,0 @@
-
-node index.js
-
-# test
-node test.js
