+++ src/index.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import TextEditor from "./TextEditor.js";
+
+let text_editor = new TextEditor();
