+++ test/test.js
@@ -1,4 +1,4 @@
-import TextEditor from "./index.js";
+import TextEditor from "../src/TextEditor.js";
 
 const editor = new TextEditor();
 //editor.openfile("CSC302.txt", "r");
