+++ test/testDav.js
@@ -1,7 +1,7 @@
 import termKit from 'terminal-kit';
-var term = termKit.terminal ;
+var term = termKit.terminal;
 
-import TextEditor from "./index.js";
+import TextEditor from "../src/TextEditor.js";
 
 const editor = new TextEditor();
 editor.init("CSC302.txt");
